Continuing the development of the profile view and the input form for updating the profile.

+++ FootballStats2022/Base/Models/PlayerEntity+Extensions.swift
@@ -47,7 +47,7 @@ extension PlayerEntity {
             Int(height_)
         }
         set {
-            height_ = Int16(newValue)
+            height_ = Int32(newValue)
         }
     }
